interpret_cnn/gradcam_app/requirements.txt
Flask==2.3.3
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
opencv-python>=4.5.0